Double-typed values

+++ Lab3/my_qsort.h
@@ -2,4 +2,4 @@
 
 void my_qsort(int* a, int left, int right);
 
-void my_qsort_bounded(int* a, int* b, int left, int right);
+void my_qsort_bounded(int* a, double* b, int left, int right);
